Activation utils cleanup (#879)
* merged activation utils

* fixed comment
